+++ test/unit/elastic_db_admin/list_repos.py
@@ -35,11 +35,11 @@ import version
 __version__ = version.__version__
 
 
-class ElasticCluster(object):
+class ElasticSearch(object):
 
-    """Class:  ElasticCluster
+    """Class:  ElasticSearch
 
-    Description:  Class representation of the ElasticCluster class.
+    Description:  Class representation of the ElasticSearch class.
 
     Methods:
         __init__ -> Initialize configuration environment.
@@ -81,7 +81,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.es = ElasticCluster()
+        self.es = ElasticSearch()
 
     @mock.patch("elastic_db_admin.elastic_libs.list_repos2")
     def test_list_repos(self, mock_lib):