+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Breaking Change
 
 ### Changed
+- elastic.py.TEMPLATE:  Changed host variable to cluster setting.
 - run_program:  Replaced ElasticCluster with ElasticSearch class instance call.