+++ CHANGELOG.md
@@ -4,6 +4,30 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [0.2.6] - 2019-11-15
+- Breaking Change
+
+### Added
+- print_failures:  Print the failed dumps in the current repository.
+- print_dumps:  Print the dumps in the current repository.
+
+### Changed
+- failed_dumps:  Refactored the function to print failed dumps in current repository or all repositories.
+- list_dumps:  Refactored the function to print dumps in current repository or all repositories.
+- check_status:  Removed -j option.
+- \_process_data:  Removed -j option.
+- get_status:  Removed -j option.
+- \_get_data:  Removed -j option.
+- get_status:  Replaced ElasticStatus with ElasticSearchStatus class instance call.
+- check_status:  Replaced ElasticStatus with ElasticSearchStatus class instance call.
+- run_program:  Replaced ElasticCluster with ElasticSearch class instance call.
+- Documentation update.
+- elastic.py.TEMPLATE:  Changed host variable to cluster setting.
+
+### Removed
+- Removed -j option (JSON formatting) as it's the only format available when checking.
+
+
 ## [0.2.5] - 2019-09-12
 ### Fixed
 - \_get_data:  Added throwaway variables for gen_libs.merge_two_dicts call.